+++ CHANGES.rst
@@ -10,6 +10,7 @@ Note: updated docs now found at `<http://refl1d.readthedocs.org>`_
 * use dQ from datafile rather than calculating it from dL and dT
 * adjust dQ with sample broadening rather than calculating it from dL and dT
 * fix json output so saving the fit doesn't raise and error
+* fix functional magnetic profile so [rho, rho] is not treated as rho, theta.
 * wxpython is now an optional dependency
 
 2020-02-18 v0.8.10